Feature/saldo

+++ src/main/java/ru/division/of/expenses/app/models/PartitialPayer.java
@@ -1,11 +1,13 @@
 package ru.division.of.expenses.app.models;
 
 import lombok.Data;
+import lombok.ToString;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import java.math.BigDecimal;
 
 @Entity
 @Data
@@ -16,7 +18,7 @@ public class PartitialPayer extends AbstractEntity {
     private User user;
 
     @Column
-    private Double coefficient;
+    private BigDecimal coefficient;
 
     @ManyToOne
     @JoinColumn(name = "expense_id")
